refactor(driver): handle panicking
