+++ compio-driver/src/sys/iocp/cp/multi.rs
@@ -22,6 +22,7 @@ impl Port {
         self.port.attach(fd)
     }
 
+    #[allow(dead_code)]
     pub fn post(&self, res: io::Result<usize>, optr: *mut Overlapped) -> io::Result<()> {
         self.port.post(res, optr)
     }
